pyproject.toml
[tool.poetry]
name = "LLM Bot"
version = "0.1"
description = "Chat Bot for answering questions about company details"
readme = "README.md"
authors = [
    "Andrii Hladkyi <andrii.hladkyi.teacher@gmail.com>",
]
license = ""

[tool.poetry.dependencies]
python = "^3.11"
llama_index = "0.7.14"
setuptools = "^68.0.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"
mypy = "^1.4.1"
debugpy = "^1.6.7"
fastapi = "^0.99.0"
types-pyyaml = "^6.0.12.10"
openai = "^0.27.8"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
chromadb = "^0.4.3"
tiktoken = "^0.4.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
